Upload RH project
Upload project for Admin RH, made with Quarkus and works wit the React app rh-app.

+++ Quarkus/Inventario/inventario/src/main/java/com/ari/group/servicio/ProductoServicio.java
@@ -48,7 +48,7 @@ public class ProductoServicio {
 
     @Transactional
     public void eliminarProductoPorId(Long idProducto) {
-        Producto producto = productoRepositorio.findById((long) idProducto);
+        Producto producto = productoRepositorio.findById(idProducto);
         if (producto != null) {
             productoRepositorio.delete(producto);
         }
